🍱 Add:	본인이 쓴 댓글이 아니면 삭제 안되는 로직 추가(#85)

+++ src/main/kotlin/com/eatsfinder/domain/comment/service/CommentServiceImpl.kt
@@ -5,12 +5,14 @@ import com.eatsfinder.domain.comment.dto.CommentResponse
 import com.eatsfinder.domain.comment.dto.CommentResponse.Companion.from
 import com.eatsfinder.domain.comment.model.Comment
 import com.eatsfinder.domain.comment.repository.CommentRepository
+import com.eatsfinder.domain.like.repository.CommentLikeRepository
 import com.eatsfinder.domain.post.repository.PostRepository
 import com.eatsfinder.domain.user.model.MyActiveType
 import com.eatsfinder.domain.user.model.UserLog
 import com.eatsfinder.domain.user.repository.UserLogRepository
 import com.eatsfinder.domain.user.repository.UserRepository
 import com.eatsfinder.global.exception.ModelNotFoundException
+import com.eatsfinder.global.exception.profile.ImmutableUserException
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
@@ -63,6 +65,10 @@ class CommentServiceImpl(
             "comment",
             "이 댓글(${commentId})은 존재하지 않습니다."
         )
+        if (comment.userId.id != userId) {
+            throw ImmutableUserException("이 댓글을 삭제할 권한이 없습니다.")
+        }
+
         commentRepository.delete(comment)
         val userLog = userLogRepository.findUserLogByCommentIdAndMyActiveType(comment, MyActiveType.COMMENT)
             ?: throw ModelNotFoundException("userLog")
